Conflicts:
      Library/Homebrew/cmd/update-report.rb
      Library/Homebrew/dev-cmd/man.rb
      Library/Homebrew/formula_installer.rb
      Library/Homebrew/manpages/brew.1.md.erb
      Library/Homebrew/test/cleanup_spec.rb
      README.md
      docs/Manpage.md
      manpages/brew.1

+++ Library/Homebrew/cleanup.rb
@@ -1,39 +1,196 @@
 require "utils/bottles"
 require "formula"
+require "hbc/cask_loader"
+require "set"
+
+module CleanupRefinement
+  LATEST_CASK_DAYS = 7
+
+  refine Enumerator do
+    def parallel
+      queue = Queue.new
+
+      each do |element|
+        queue.enq(element)
+      end
+
+      workers = (0...Hardware::CPU.cores).map do
+        Thread.new do
+          Kernel.loop do
+            begin
+              yield queue.deq(true)
+            rescue ThreadError
+              break # if queue is empty
+            end
+          end
+        end
+      end
+
+      workers.each(&:join)
+    end
+  end
+
+  refine Pathname do
+    def incomplete?
+      extname.end_with?(".incomplete")
+    end
+
+    def nested_cache?
+      directory? && ["glide_home", "java_cache", "npm_cache"].include?(basename.to_s)
+    end
+
+    def prune?(days)
+      return false unless days
+      return true if days.zero?
+
+      # TODO: Replace with ActiveSupport's `.days.ago`.
+      mtime < ((@time ||= Time.now) - days * 60 * 60 * 24)
+    end
+
+    def stale?(scrub = false)
+      return false unless file?
+
+      stale_formula?(scrub) || stale_cask?(scrub)
+    end
+
+    private
+
+    def stale_formula?(scrub)
+      return false unless HOMEBREW_CELLAR.directory?
+
+      version = if to_s.match?(Pathname::BOTTLE_EXTNAME_RX)
+        begin
+          Utils::Bottles.resolve_version(self)
+        rescue
+          nil
+        end
+      end
+
+      version ||= basename.to_s[/\A.*(?:\-\-.*?)*\-\-(.*?)#{Regexp.escape(extname)}\Z/, 1]
+      version ||= basename.to_s[/\A.*\-\-?(.*?)#{Regexp.escape(extname)}\Z/, 1]
+
+      return false unless version
+
+      version = Version.new(version)
+
+      return false unless formula_name = basename.to_s[/\A(.*?)(?:\-\-.*?)*\-\-?(?:#{Regexp.escape(version)})/, 1]
+
+      formula = begin
+        Formulary.from_rack(HOMEBREW_CELLAR/formula_name)
+      rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
+        return false
+      end
+
+      resource_name = basename.to_s[/\A.*?\-\-(.*?)\-\-?(?:#{Regexp.escape(version)})/, 1]
+
+      if resource_name == "patch"
+        patch_hashes = formula.stable&.patches&.select(&:external?)&.map(&:resource)&.map(&:version)
+        return true unless patch_hashes&.include?(Checksum.new(:sha256, version.to_s))
+      elsif resource_name && resource_version = formula.stable&.resources&.dig(resource_name)&.version
+        return true if resource_version != version
+      elsif version.is_a?(PkgVersion)
+        return true if formula.pkg_version > version
+      elsif formula.version > version
+        return true
+      end
+
+      return true if scrub && !formula.installed?
+
+      return true if Utils::Bottles.file_outdated?(formula, self)
+
+      false
+    end
+
+    def stale_cask?(scrub)
+      return false unless name = basename.to_s[/\A(.*?)\-\-/, 1]
+
+      cask = begin
+        Hbc::CaskLoader.load(name)
+      rescue Hbc::CaskUnavailableError
+        return false
+      end
+
+      unless basename.to_s.match?(/\A#{Regexp.escape(name)}\-\-#{Regexp.escape(cask.version)}\b/)
+        return true
+      end
+
+      return true if scrub && !cask.versions.include?(cask.version)
+
+      if cask.version.latest?
+        # TODO: Replace with ActiveSupport's `.days.ago`.
+        return mtime < ((@time ||= Time.now) - LATEST_CASK_DAYS * 60 * 60 * 24)
+      end
+
+      false
+    end
+  end
+end
+
+using CleanupRefinement
 
 module Homebrew
-  module Cleanup
-    @disk_cleanup_size = 0
+  class Cleanup
+    extend Predicable
 
-    class << self
-      attr_reader :disk_cleanup_size
+    attr_predicate :dry_run?, :scrub?
+    attr_reader :args, :days, :cache
+    attr_reader :disk_cleanup_size
+
+    def initialize(*args, dry_run: false, scrub: false, days: nil, cache: HOMEBREW_CACHE)
+      @disk_cleanup_size = 0
+      @args = args
+      @dry_run = dry_run
+      @scrub = scrub
+      @days = days
+      @cache = cache
+      @cleaned_up_paths = Set.new
     end
 
-    module_function
+    def clean!
+      if args.empty?
+        cleanup_lockfiles
+        Formula.installed.sort_by(&:name).each do |formula|
+          cleanup_formula(formula)
+        end
+        cleanup_cache
+        cleanup_logs
+        return if dry_run?
+        rm_ds_store
+      else
+        args.each do |arg|
+          formula = begin
+            Formula[arg]
+          rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
+            nil
+          end
 
-    def cleanup
-      cleanup_cellar
-      cleanup_cache
-      cleanup_logs
-      return if ARGV.dry_run?
-      cleanup_lockfiles
-      rm_ds_store
-    end
+          cask = begin
+            Hbc::CaskLoader.load(arg)
+          rescue Hbc::CaskUnavailableError
+            nil
+          end
 
-    def update_disk_cleanup_size(path_size)
-      @disk_cleanup_size += path_size
+          cleanup_formula(formula) if formula
+          cleanup_cask(cask) if cask
+        end
+      end
     end
 
     def unremovable_kegs
       @unremovable_kegs ||= []
     end
 
-    def cleanup_cellar(formulae = Formula.installed)
-      formulae.each(&method(:cleanup_formula))
-    end
-
     def cleanup_formula(formula)
       formula.eligible_kegs_for_cleanup.each(&method(:cleanup_keg))
+      cleanup_cache(Pathname.glob(cache/"#{formula.name}--*"))
+      rm_ds_store([formula.rack])
+      cleanup_lockfiles(FormulaLock.new(formula.name).path)
+    end
+
+    def cleanup_cask(cask)
+      cleanup_cache(Pathname.glob(cache/"Cask/#{cask.token}--*"))
+      rm_ds_store([cask.caskroom_path])
+      cleanup_lockfiles(CaskLock.new(cask.token).path)
     end
 
     def cleanup_keg(keg)
@@ -43,25 +200,23 @@ module Homebrew
       unremovable_kegs << keg
     end
 
+    DEFAULT_LOG_DAYS = 14
+
     def cleanup_logs
       return unless HOMEBREW_LOGS.directory?
       HOMEBREW_LOGS.subdirs.each do |dir|
-        cleanup_path(dir) { dir.rmtree } if prune?(dir, days_default: 14)
+        cleanup_path(dir) { dir.rmtree } if dir.prune?(days || DEFAULT_LOG_DAYS)
       end
     end
 
-    def cleanup_cache(cache = HOMEBREW_CACHE)
-      return unless cache.directory?
-      cache.children.each do |path|
-        if path.to_s.end_with? ".incomplete"
-          cleanup_path(path) { path.unlink }
-          next
-        end
-        if %w[glide_home java_cache npm_cache].include?(path.basename.to_s) && path.directory?
-          cleanup_path(path) { FileUtils.rm_rf path }
-          next
-        end
-        if prune?(path)
+    def cleanup_cache(entries = nil)
+      entries ||= [cache, cache/"Cask"].select(&:directory?).flat_map(&:children)
+
+      entries.each do |path|
+        next cleanup_path(path) { path.unlink } if path.incomplete?
+        next cleanup_path(path) { FileUtils.rm_rf path } if path.nested_cache?
+
+        if path.prune?(days)
           if path.file?
             cleanup_path(path) { path.unlink }
           elsif path.directory? && path.to_s.include?("--")
@@ -70,99 +225,51 @@ module Homebrew
           next
         end
 
-        next unless path.file?
-        file = path
-
-        if file.to_s =~ Pathname::BOTTLE_EXTNAME_RX
-          version = begin
-                      Utils::Bottles.resolve_version(file)
-                    rescue
-                      file.version
-                    end
-        else
-          version = file.version
-        end
-        next unless version
-        next unless (name = file.basename.to_s[/\A(.*?)\-\-?(?:#{Regexp.escape(version)})/, 1])
-
-        next unless HOMEBREW_CELLAR.directory?
-
-        begin
-          f = Formulary.from_rack(HOMEBREW_CELLAR/name)
-        rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
-          next
-        end
-
-        file_is_stale = if version.is_a?(PkgVersion)
-          f.pkg_version > version
-        else
-          f.version > version
-        end
-
-        if file_is_stale || ARGV.switch?("s") && !f.installed? || Utils::Bottles.file_outdated?(f, file)
-          cleanup_path(file) { file.unlink }
-        end
+        next cleanup_path(path) { path.unlink } if path.stale?(scrub?)
       end
     end
 
     def cleanup_path(path)
-      if ARGV.dry_run?
+      return unless @cleaned_up_paths.add?(path)
+
+      disk_usage = path.disk_usage
+
+      if dry_run?
         puts "Would remove: #{path} (#{path.abv})"
       else
         puts "Removing: #{path}... (#{path.abv})"
         yield
       end
 
-      update_disk_cleanup_size(path.disk_usage)
+      @disk_cleanup_size += disk_usage
     end
 
-    def cleanup_lockfiles
+    def cleanup_lockfiles(*lockfiles)
       return unless HOMEBREW_LOCK_DIR.directory?
-      candidates = HOMEBREW_LOCK_DIR.children
-      lockfiles  = candidates.select(&:file?)
+
+      if lockfiles.empty?
+        lockfiles = HOMEBREW_LOCK_DIR.children.select(&:file?)
+      end
+
       lockfiles.each do |file|
         next unless file.readable?
-        file.open(File::RDWR).flock(File::LOCK_EX | File::LOCK_NB) && file.unlink
-      end
-    end
+        next unless file.open(File::RDWR).flock(File::LOCK_EX | File::LOCK_NB)
 
-    def rm_ds_store
-      paths = Queue.new
-      %w[Cellar Frameworks Library bin etc include lib opt sbin share var]
-        .map { |p| HOMEBREW_PREFIX/p }.each { |p| paths << p if p.exist? }
-      workers = (0...Hardware::CPU.cores).map do
-        Thread.new do
-          Kernel.loop do
-            begin
-              quiet_system "find", paths.deq(true), "-name", ".DS_Store", "-delete"
-            rescue ThreadError
-              break # if queue is empty
-            end
-          end
+        begin
+          cleanup_path(file) { file.unlink }
+        ensure
+          file.open(File::RDWR).flock(File::LOCK_UN) if file.exist?
         end
       end
-      workers.map(&:join)
     end
 
-    def prune?(path, options = {})
-      @time ||= Time.now
+    def rm_ds_store(dirs = nil)
+      dirs ||= %w[Caskroom Cellar Frameworks Library bin etc include lib opt sbin share var]
+               .map { |path| HOMEBREW_PREFIX/path }
 
-      path_modified_time = path.mtime
-      days_default = options[:days_default]
-
-      prune = ARGV.value "prune"
-
-      return true if prune == "all"
-
-      prune_time = if prune
-        @time - 60 * 60 * 24 * prune.to_i
-      elsif days_default
-        @time - 60 * 60 * 24 * days_default.to_i
+      dirs.select(&:directory?).each.parallel do |dir|
+        system_command "find", args: [dir, "-name", ".DS_Store", "-delete"], print_stderr: false
       end
-
-      return false unless prune_time
-
-      path_modified_time < prune_time
     end
   end
 end
